pyproject.toml
[tool.poetry]
name = "steering-wizard"
version = "0.1.0"
description = "A Python CLI wizard for creating standardized Kiro steering documents"
authors = ["Developer <developer@example.com>"]
readme = "README.md"
packages = [{include = "steering_wizard", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.0.0"
rich = "^13.0.0"
pyyaml = "^6.0.0"
jinja2 = "^3.1.0"
jsonschema = "^4.19.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
hypothesis = "^6.88.0"
black = "^23.0.0"
pylint = "^3.0.0"
mypy = "^1.6.0"
pytest-cov = "^4.1.0"

[tool.poetry.scripts]
steering-wizard = "steering_wizard.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pylint.messages_control]
disable = [
    "C0114",  # missing-module-docstring
    "C0115",  # missing-class-docstring
    "C0116",  # missing-function-docstring
    "R0903",  # too-few-public-methods
    "W0613",  # unused-argument
]

[tool.pylint.format]
max-line-length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false